+++ validator/libOpenAPI/fileGetter.js
@@ -1,7 +1,14 @@
-var XMLHttpRequest = require("xmlhttprequest").XMLHttpRequest;
+/*
+This class has the objective of doing http requests for getting external objects
+@author Francisco F. Cardoso | T-TALK
+*/
 
+var XMLHttpRequest = require("xmlhttprequest").XMLHttpRequest;
 var results;
 
+/**
+ * This method clears all objects before validating the next file
+ */
 exports.clear = function() {    
     completed_requests = 0;
     results = {
@@ -11,6 +18,10 @@ exports.clear = function() {
     };
 }
 
+/**
+ * This methods runs through the list of schemas and returns a list of schemas 
+ * @param {*} schemaObjList Array
+ */
 exports.getAllExternalFiles = function (schemaObjList) {
     var schemaUrlList = getSchemaUrls(schemaObjList);  
     for (var i in schemaUrlList) {
@@ -19,6 +30,10 @@ exports.getAllExternalFiles = function (schemaObjList) {
     return results;
 }
 
+/**
+ *  This methods returns all schemaUrls from schemaObjList
+ * @param {*} schemaObjList Array
+ */
 var getSchemaUrls = function (schemaObjList) { 
     var schemaUrlList = [];
     for (var i in schemaObjList) {
@@ -29,6 +44,11 @@ var getSchemaUrls = function (schemaObjList) {
     return schemaUrlList;
 }
 
+/**
+ * This method does the http get request and place the answer in an array
+ * If the reference returns a 404 (or any other error) will be placed in a notFoundSchemas list
+ * @param {*} url 
+ */
 var getExternalFile = function (url) {    
     var rawFile = new XMLHttpRequest();
     rawFile.open("GET", url, false);
