+++ validator/test/testOpenAPI.js
@@ -1,12 +1,16 @@
 "use strict";
+/*
+This class contains MOCHA test cases for OpenAPI files
+@author Francisco F. Cardoso | T-TALK
+*/
 
 var expect = require('expect.js');
 var fs = require('fs');
 var path = require('path');
-var pathValidator = require('../lib/pathValidator.js');
-var jsonHandler = require('../lib/jsonHandler.js');
-var fileGetter = require('../lib/fileGetter.js');
-var schemaReferenceFromApi = require('../lib/schemaReferenceFromApiValidator.js');
+var pathValidator = require('../libOpenAPI/pathValidator.js');
+var jsonHandler = require('../libOpenAPI/jsonHandler.js');
+var fileGetter = require('../libOpenAPI/fileGetter.js');
+var schemaReferenceFromApi = require('../libOpenAPI/schemaReferenceFromApiValidator.js');
 
 var expect = require('chai').expect;
 
@@ -143,7 +147,7 @@ fs.readdir(dirname, function (err, filenames) {
           it("should contain the same Id property name in URL and body", function () {
             var errorMessage = "";
             if (schemaReferenceFromApiResult.erroredPath)
-              errorMessage = "Check the endpoint '" + schemaReferenceFromApiResult.erroredPath + "'";
+              errorMessage = "Check the endpoint '" + schemaReferenceFromApiResult.erroredPath + "'. It may be a typo or canse sensitve difference";
             if (schemaReferenceFromApiResult.validObject)
               expect(schemaReferenceFromApiResult.containsTheSameKeyInUrlAndBody, errorMessage).to.be.true;
           });
